split into two tickmarks components
